Added hitWindow to replay, and moved HitWindow to HitStat
