+++ Main/include/Game.hpp
@@ -4,6 +4,7 @@
 #include <Beatmap/MapDatabase.hpp>
 #include "json.hpp"
 #include "GameFailCondition.hpp"
+#include "HitStat.hpp"
 
 class MultiplayerScreen;
 
@@ -31,6 +32,8 @@ struct ScoreReplay
 	int32 maxScore = 0;
 	size_t nextHitStat = 0;
 	Vector<SimpleHitStat> replay;
+
+	HitWindow hitWindow = HitWindow::NORMAL;
 };
 
 GameFlags operator|(const GameFlags& a, const GameFlags& b);
