+++ Main/src/HitStat.cpp
@@ -1,5 +1,9 @@
 #include "stdafx.h"
 #include "HitStat.hpp"
+#include "GameConfig.hpp"
+
+const HitWindow HitWindow::NORMAL = HitWindow(46, 92);
+const HitWindow HitWindow::HARD = HitWindow(23, 46);
 
 HitStat::HitStat(ObjectState* object) : object(object)
 {
@@ -8,4 +12,42 @@ HitStat::HitStat(ObjectState* object) : object(object)
 bool HitStat::operator<(const HitStat& other)
 {
 	return time < other.time;
+}
+
+HitWindow HitWindow::FromConfig()
+{
+	HitWindow hitWindow = HitWindow(
+		g_gameConfig.GetInt(GameConfigKeys::HitWindowPerfect),
+		g_gameConfig.GetInt(GameConfigKeys::HitWindowGood),
+		g_gameConfig.GetInt(GameConfigKeys::HitWindowHold)
+	);
+
+	if (!(hitWindow <= HitWindow::NORMAL))
+	{
+		Log("HitWindow is automatically adjusted to NORMAL", Logger::Severity::Warning);
+		hitWindow = HitWindow::NORMAL;
+
+		g_gameConfig.Set(GameConfigKeys::HitWindowPerfect, hitWindow.perfect);
+		g_gameConfig.Set(GameConfigKeys::HitWindowGood, hitWindow.good);
+		g_gameConfig.Set(GameConfigKeys::HitWindowHold, hitWindow.hold);
+	}
+
+	return hitWindow;
+}
+
+void HitWindow::ToLuaTable(lua_State* L)
+{
+	auto pushIntToTable = [&](const char* name, int data)
+	{
+		lua_pushstring(L, name);
+		lua_pushinteger(L, data);
+		lua_settable(L, -3);
+	};
+
+	lua_newtable(L);
+	pushIntToTable("type", static_cast<int>(GetType()));
+	pushIntToTable("perfect", perfect);
+	pushIntToTable("good", good);
+	pushIntToTable("hold", hold);
+	pushIntToTable("miss", miss);
 }
