+++ Main/src/Scoring.cpp
@@ -4,15 +4,6 @@
 #include <math.h>
 #include "GameConfig.hpp"
 
-extern "C"
-{
-#include "lua.h"
-#include "lauxlib.h"
-}
-
-const HitWindow HitWindow::NORMAL = HitWindow(46, 92);
-const HitWindow HitWindow::HARD = HitWindow(23, 46);
-
 const float Scoring::idleLaserSpeed = 1.0f;
 
 Scoring::Scoring()
@@ -1426,42 +1417,4 @@ TickFlags operator|(const TickFlags& a, const TickFlags& b)
 TickFlags operator&(const TickFlags& a, const TickFlags& b)
 {
 	return (TickFlags)((uint8)a & (uint8)b);
-}
-
-HitWindow HitWindow::FromConfig()
-{
-	HitWindow hitWindow = HitWindow(
-		g_gameConfig.GetInt(GameConfigKeys::HitWindowPerfect),
-		g_gameConfig.GetInt(GameConfigKeys::HitWindowGood),
-		g_gameConfig.GetInt(GameConfigKeys::HitWindowHold)
-	);
-
-	if (!(hitWindow <= HitWindow::NORMAL))
-	{
-		Log("HitWindow is automatically adjusted to NORMAL", Logger::Severity::Warning);
-		hitWindow = HitWindow::NORMAL;
-
-		g_gameConfig.Set(GameConfigKeys::HitWindowPerfect, hitWindow.perfect);
-		g_gameConfig.Set(GameConfigKeys::HitWindowGood, hitWindow.good);
-		g_gameConfig.Set(GameConfigKeys::HitWindowHold, hitWindow.hold);
-	}
-
-	return hitWindow;
-}
-
-void HitWindow::ToLuaTable(lua_State* L)
-{
-	auto pushIntToTable = [&](const char* name, int data)
-	{
-		lua_pushstring(L, name);
-		lua_pushinteger(L, data);
-		lua_settable(L, -3);
-	};
-
-	lua_newtable(L);
-	pushIntToTable("type", static_cast<int>(GetType()));
-	pushIntToTable("perfect", perfect);
-	pushIntToTable("good", good);
-	pushIntToTable("hold", hold);
-	pushIntToTable("miss", miss);
 }
